requirements.txt
amqp==5.2.0
ansicon==1.89.0
arrow==1.3.0
asgiref==3.7.2
attrs==23.2.0
billiard==4.2.0
blessed==1.20.0
celery==5.4.0
certifi==2024.2.2
cffi
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.5
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-picklefield==3.1
django-redis==5.4.0
django-shortuuidfield==0.1.3
djangorestframework==3.15.1
dnspython==2.6.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.4.1
et-xmlfile==1.1.0
eventlet==0.35.2
greenlet
idna==3.6
inflection==0.5.1
jinxed==1.2.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
mysqlclient
numpy
openpyxl==3.1.2
pandas
pillow
prompt-toolkit==3.0.43
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.20.0
PyJWT==2.8.0
PyMySQL==1.1.0
pyOpenSSL==24.0.0
pypinyin==0.51.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
referencing==0.34.0
requests==2.31.0
rpds-py
setuptools==69.2.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
zope.event==5.0
zope.interface==6.2
gevent
